Rename calingen.forms.tex to calingen.forms.generation
